added the gala, fundraising days, and fundraising guide pages
It should be all good, however there were a lot of changes and
rearrangements so it’ll be necessary to double check the links. I’ve
done it multiple times but it’s easy to miss stuff.

Also, the fundraising guide page under the Fundraising tab is missing
information for “Templates” and I don’t have the “fundraising guide”
PDF, so that still needs to be linked.